requirements.txt
python-telegram-bot==20.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
google-genai==1.7.0
pandas==2.1.3
openpyxl==3.1.2
gspread==5.12.0
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9